treewide: Move public Rekor data structures to top-level `rekor` module
